feat: erged models and projects
